code qui marque des points

+++ embedded_program/include/sensors.h
@@ -1,6 +1,6 @@
 #ifndef __SENSORS_H__
 #define __SENSORS_H__
 
-int readUltrasonic();
+float readUltrasonic();
 
 #endif
